scripts/aminstall.py: remove demo script
The script was started but never finished, as it was not useful
for myself since I can do manual installation anyway, and possi-
ble users are not many.

Especially considering this is an official demo script -- which
means it should be refered to by the distro implementers to
write their installers, it's better deleted than left incomplete,
when the downstream implementer should refer to the doc for a
better understanding of ampart's API and write their own scripts
in a way they like, rather than adapting a script from here, which
will make their script limited to my demo script's style.

If you're a user, you're probably reading this when wanting to
complain why aminstall.sh (note it's not the .py version here,
the .sh version is never maintained here)  was removed from
EmuELEC: it's removed because I don't have time to update it
while the API was changed a lot, which leaves the choice to be
either 1) old ampart + aminstall, or 2) new ampart + no aminstall,
or 3) no ampart at all. This is written before the actual change
was made to the EmuELEC repo but mostly it would be removed.

For a personal note, this should probably be my last commit on
ampart for a long time, I have an exam at the end of the year
and should not waste more time on non-study business if I don't
want to fail at the post-graduate admission for the third time
and disapppoint my families for another time. You should be
reading this message if you're wondering why ampart wasn't
updated for a long time: be patient, I'll probably return next
year, at January.